+++ Common/src/main/java/net/createmod/ponder/mixin/client/accessor/GameRendererAccessor.java
@@ -8,6 +8,7 @@ import net.minecraft.client.renderer.GameRenderer;
 
 @Mixin(GameRenderer.class)
 public interface GameRendererAccessor {
+	// MC 1.21.2: getFov now returns float instead of double
 	@Invoker("getFov")
-	double catnip$callGetFov(Camera camera, float partialTicks, boolean useFOVSetting);
+	float catnip$callGetFov(Camera camera, float partialTicks, boolean useFOVSetting);
 }
